+++ wkcuber/wkcuber/upload.py
@@ -5,6 +5,7 @@ from os import environ
 
 from webknossos import webknossos_context
 from webknossos.client._defaults import DEFAULT_WEBKNOSSOS_URL
+from webknossos.client._upload_dataset import DEFAULT_SIMULTANEOUS_UPLOADS
 from wkcuber.api.dataset import Dataset
 
 from .utils import add_verbose_flag, setup_logging
@@ -30,9 +31,15 @@ def create_parser() -> ArgumentParser:
     parser.add_argument(
         "--jobs",
         "-j",
-        default=5,
+        default=DEFAULT_SIMULTANEOUS_UPLOADS,
         type=int,
-        help="Number of simultaneous upload processes.",
+        help=f"Number of simultaneous upload processes. Defaults to {DEFAULT_SIMULTANEOUS_UPLOADS}.",
+    )
+
+    parser.add_argument(
+        "--name",
+        help="Specify a new name for the dataset. Defaults to the name specified in `datasource-properties.json`.",
+        default=None,
     )
 
     add_verbose_flag(parser)
@@ -41,13 +48,10 @@ def create_parser() -> ArgumentParser:
 
 
 def upload_dataset(
-    source_path: Path,
-    url: str,
-    token: Optional[str],
-    jobs: int,
+    source_path: Path, url: str, token: Optional[str], jobs: int, name: Optional[str]
 ) -> None:
     with webknossos_context(url=url, token=token):
-        Dataset.open(source_path).upload(jobs)
+        Dataset.open(source_path).upload(new_dataset_name=name, jobs=jobs)
 
 
 if __name__ == "__main__":
@@ -63,4 +67,4 @@ if __name__ == "__main__":
         token is not None
     ), f"An auth token needs to be supplied either through the --token command line arg or the WK_TOKEN environment variable. Retrieve your auth token on {url}/auth/token."
 
-    upload_dataset(args.source_path, url, token, args.jobs)
+    upload_dataset(args.source_path, url, token, args.jobs, args.name)