+++ webknossos/Changelog.md
@@ -12,6 +12,8 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 ### Breaking Changes
 
 ### Added
+- Added `Annotation.save("file_name")` to save an annotation to a file and `Annotation.temporary_volume_annotation_layer_copy()` to read from the volume layer of an annotation as a WK dataset. [#528](https://github.com/scalableminds/webknossos-libs/pull/528)
+- Added `layers_to_link` parameter to `Dataset.upload()` so that layers don't need to be uploaded again if they already exist in another dataset on webKnossos. [#544](https://github.com/scalableminds/webknossos-libs/pull/544)
 
 ### Changed
 
@@ -34,7 +36,6 @@ For upgrade instructions, please check the respective *Breaking Changes* section
   * `Skeleton.write()` → `Skeleton.save()`
   The deprecated methods will be removed in future releases.
   [#520](https://github.com/scalableminds/webknossos-libs/pull/520)
-
 
 ### Changed
 - The detailed output of e.g. downsampling was replaced with a progress bar. [#527](https://github.com/scalableminds/webknossos-libs/pull/527)